Price Splitter
This was the final project for my visual programming class. It's a
window form program coded in visual basic.
This program allows a user to enter in a money amount and number of
people/items and display the evenly split amount. Tips, taxes, and
shipping are taken into account depending on which form is used.